Revert "pkg/release: move from pkg/framework/resource (#137)"
This reverts commit 4c120076417c76d2f556f8081d18a0106d5c312b.
