+++ pkg/framework/resource/resource.go
@@ -1,4 +1,4 @@
-package release
+package resource
 
 import (
 	"context"
